requirements.txt
bcrypt
fastapi==0.109.2
pydantic==2.6.1
PyJWT==2.8.0
python-dotenv==1.0.1
typing_extensions==4.9.0
uvicorn==0.27.1
python-multipart